use only usize

+++ apollo-router/src/configuration/mod.rs
@@ -461,7 +461,7 @@ pub fn validate_configuration(raw_yaml: &str) -> Result<(), ConfigurationError> 
                     .enumerate()
                     .filter_map(|(idx, e)| {
                         if let Some(element) = yaml.get_element(&e.instance_path) {
-                            const NUMBER_OF_PREVIOUS_LINES_TO_DISPLAY: isize = 5;
+                            const NUMBER_OF_PREVIOUS_LINES_TO_DISPLAY: usize = 5;
                             match element {
                                 yaml::Value::String(value, marker) => {
                                     // Dirty hack.
@@ -477,11 +477,8 @@ pub fn validate_configuration(raw_yaml: &str) -> Result<(), ConfigurationError> 
                                         return None;
                                     }
 
-                                    let lines = yaml_split_by_lines[0.max(
-                                        marker.line() as isize
-                                            - NUMBER_OF_PREVIOUS_LINES_TO_DISPLAY,
-                                    )
-                                        as usize
+                                    let lines = yaml_split_by_lines[0
+                                        .max(marker.line() - NUMBER_OF_PREVIOUS_LINES_TO_DISPLAY)
                                         ..marker.line()]
                                         .iter()
                                         .join("\n");
@@ -499,9 +496,8 @@ pub fn validate_configuration(raw_yaml: &str) -> Result<(), ConfigurationError> 
                                     let (start_marker, end_marker) = (m, seq_element.end_marker());
 
                                     let offset = 0.max(
-                                        start_marker.line() as isize
-                                            - NUMBER_OF_PREVIOUS_LINES_TO_DISPLAY,
-                                    ) as usize;
+                                        start_marker.line() - NUMBER_OF_PREVIOUS_LINES_TO_DISPLAY,
+                                    );
                                     let lines = yaml_split_by_lines[offset..end_marker.line()]
                                         .iter()
                                         .enumerate()
@@ -529,9 +525,8 @@ pub fn validate_configuration(raw_yaml: &str) -> Result<(), ConfigurationError> 
                                         map_value.end_marker(),
                                     );
                                     let offset = 0.max(
-                                        start_marker.line() as isize
-                                            - NUMBER_OF_PREVIOUS_LINES_TO_DISPLAY,
-                                    ) as usize;
+                                        start_marker.line() - NUMBER_OF_PREVIOUS_LINES_TO_DISPLAY,
+                                    );
                                     let lines = yaml_split_by_lines[offset..end_marker.line()]
                                         .iter()
                                         .enumerate()
